pyproject.toml
[project]
name = "test-remote-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastmcp>=3.0.2",
    "python-dotenv>=1.2.1",
]
